knowledge_web: extract from ideas.py; participation is emergent, no registry
The cross-CA layer had accreted inside the module for one CA type, so
every consumer imported the type table from 'ideas' — how a second
registry gets born one import at a time. Extracted to knowledge_web.py
with the graph->web renames the terminology assessment called for:
build_knowledge_web, query_knowledge_web, detect_web_gaps,
generate_web_html, format_web_cluster_cross_ca, format_web_tree. No
compatibility aliases; the test suite enumerated the real consumers
(cli, viz, doctor, two test modules) and each was updated properly.

The registry itself is gone, per operator direction. Participation is
emergent — writing [[links]] into an artifact IS the act of joining, and
scan locations are discovered from the filesystem rather than declared:
every subdirectory of agent/private and agent/public, personal policies,
subagent trees, and the framework policies. The participates flag, the
declared-but-not-participating entries, and the doctor's whole
registry-integrity check die with it: under universal participation
there is no registry to drift from, so the undeclared-but-real state the
acute finding existed to catch is now inexpressible rather than checked.

What remains in code is two derivation rules that execute definitions
owned by the CA-type policies (Policy-as-Spec): unit-of-node
(experiments: protocol+analysis; roadmaps: roadmap.md) and node class
(policies normative; checkpoints/roadmaps/sprints/amail temporal).
The doctor and the builder now share one walk (iter_web_files), so a
checker cannot drift from the thing it checks by keeping its own copy.

Verification: the web is IDENTICAL before and after on this corpus —
177 nodes / 5659 edges / 2744 cross-CA / 109 concepts — while the walk
now covers 247 files across 12 discovered types, including two
(delegation_trails, personal_policies) no registry had ever named.
Full suite: 1316 passed.
